Move stable classes dir to the proper test suite
